Multiple changes:
Convert OBJECT libraries to plain STATIC libraries.
Remove useless Swap.h include
Set c++ standard to c++14

Use jlsignals for signals/slots
 - All event handlers are typed now.
 - All signals are global variables.

Rename Urho2D dir to plain 2D
Sprinkle some overrides around the place.
Return QString()s instead of s_dummy in some functions.
Add move-construction and move-assign to ArrayPtr.h

+++ Source/Lutefisk3D/Audio/SoundSource3D.cpp
@@ -223,7 +223,7 @@ void SoundSource3D::CalculateAttenuation()
 
         // Distance attenuation
         if (interval > 0.0f)
-            attenuation_ = powf(1.0f - Clamp(distance - nearDistance_, 0.0f, interval) / interval, rolloffFactor_);
+            attenuation_ = std::pow(1.0f - Clamp(distance - nearDistance_, 0.0f, interval) / interval, rolloffFactor_);
         else
             attenuation_ = distance <= nearDistance_ ? 1.0f : 0.0f;
